+++ apps/e2e/ag-grid-storybook/src/app/ag-grid/ag-grid-stories.component.ts
@@ -4,6 +4,7 @@ import {
   ChangeDetectorRef,
   Component,
   Inject,
+  Input,
   OnDestroy,
   OnInit,
   ViewEncapsulation,
@@ -44,10 +45,12 @@ interface DataSet {
 export class AgGridStoriesComponent
   implements OnInit, AfterViewInit, OnDestroy
 {
+  @Input() public compact = false;
+
   public dataSets: DataSet[] = [
     {
       id: 'back-to-top',
-      data: data.slice(40, 48),
+      data: data.slice(40, 78),
     },
     {
       id: 'row-delete',