+++ apps/e2e/ag-grid-storybook/src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { Route, RouterModule } from '@angular/router';
+import { provideInitialTheme } from '@skyux/theme';
 
 import { AppComponent } from './app.component';
 
@@ -37,6 +38,7 @@ if (routes.length > 0 && routes.findIndex((r) => r.path === '') === -1) {
     BrowserModule,
     RouterModule.forRoot(routes, { initialNavigation: 'enabledBlocking' }),
   ],
+  providers: [provideInitialTheme('modern')],
   bootstrap: [AppComponent],
 })
 export class AppModule {}