+++ apps/playground/src/app/components/forms/checkbox/checkbox.component.ts
@@ -6,6 +6,7 @@ import {
   UntypedFormBuilder,
   UntypedFormGroup,
   ValidationErrors,
+  Validators,
 } from '@angular/forms';
 
 @Component({
@@ -37,7 +38,7 @@ export class CheckboxComponent implements OnInit {
 
     this.contactMethod = this.#formBuilder.group({
       email: new FormControl(false),
-      phone: new FormControl(false),
+      phone: new FormControl(false, [Validators.requiredTrue]),
       text: new FormControl(false),
     });
 
@@ -64,7 +65,7 @@ export class CheckboxComponent implements OnInit {
 
   public ngOnInit(): void {
     this.reactiveFormGroup = this.#formBuilder.group({
-      reactiveCheckbox: new FormControl(undefined),
+      reactiveCheckbox: new FormControl(undefined, [Validators.requiredTrue]),
     });
   }
 
@@ -78,5 +79,10 @@ export class CheckboxComponent implements OnInit {
 
   public toggleInlineHelp(): void {
     this.showInlineHelp = !this.showInlineHelp;
+  }
+
+  protected onSubmitContactMethod(): void {
+    this.contactMethod.markAsDirty();
+    this.contactMethod.markAsTouched();
   }
 }