+++ apps/playground/src/app/components/forms/single-file-attachment/single-file-attachment.component.ts
@@ -29,6 +29,8 @@ export class SingleFileAttachmentComponent implements OnInit {
 
   public reactiveUploadError: string;
 
+  public helpKey: string | undefined;
+
   constructor(private formBuilder: UntypedFormBuilder) {}
 
   public fileClick($event: SkyFileAttachmentClick): void {