+++ libs/components/ag-grid/src/lib/modules/ag-grid/ag-grid.service.spec.ts
@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { expect } from '@skyux-sdk/testing';
 import { SkyDateService } from '@skyux/datetime';
 import {
@@ -27,9 +27,7 @@ import {
 import { BehaviorSubject } from 'rxjs';
 
 import { SkyAgGridAdapterService } from './ag-grid-adapter.service';
-import { SkyAgGridModule } from './ag-grid.module';
 import { SkyAgGridService } from './ag-grid.service';
-import { SkyAgGridFixtureComponent } from './fixtures/ag-grid.component.fixture';
 import { SkyCellClass } from './types/cell-class';
 import { SkyCellType } from './types/cell-type';
 
@@ -95,9 +93,7 @@ describe('SkyAgGridService', () => {
         gridOptions: overrideGridOptions,
       });
 
-      expect(defaultGridOptions.headerHeight).not.toEqual(newHeight);
       expect(defaultGridOptions.rowHeight).not.toEqual(newHeight);
-      expect(mergedGridOptions.headerHeight).toEqual(newHeight);
       expect(mergedGridOptions.rowHeight).toEqual(newHeight);
     });
 
@@ -209,7 +205,7 @@ describe('SkyAgGridService', () => {
       expect(mergedCellClassRules?.[SkyCellClass.Uneditable]).toBeDefined();
     });
 
-    it('should set rowHeight and headerHeight for modern theme', () => {
+    it('should set icons for modern theme', () => {
       // Trigger change to modern theme
       mockThemeSvc.settingsChange.next({
         currentSettings: new SkyThemeSettings(
@@ -224,8 +220,6 @@ describe('SkyAgGridService', () => {
         gridOptions: {},
       });
 
-      expect(modernThemeGridOptions.rowHeight).toBe(60);
-      expect(modernThemeGridOptions.headerHeight).toBe(60);
       expect(typeof modernThemeGridOptions.icons?.['sortDescending']).toBe(
         'function',
       );
@@ -235,10 +229,7 @@ describe('SkyAgGridService', () => {
     });
 
     it('should unsubscribe from the theme service when destroyed', () => {
-      const overrideOptions = { gridOptions: {} };
-      let gridOptions = agGridService.getGridOptions(overrideOptions);
-
-      expect(gridOptions.rowHeight).toBe(38);
+      expect(agGridService.getHeaderHeight()).toBe(37);
 
       // Trigger change to modern theme
       mockThemeSvc.settingsChange.next({
@@ -251,8 +242,7 @@ describe('SkyAgGridService', () => {
       });
 
       // Get new grid options after theme change
-      gridOptions = agGridService.getGridOptions(overrideOptions);
-      expect(gridOptions.rowHeight).toBe(60);
+      expect(agGridService.getHeaderHeight()).toBe(60);
 
       // Destroy subscription
       agGridService.ngOnDestroy();
@@ -268,8 +258,7 @@ describe('SkyAgGridService', () => {
       });
 
       // Get new grid options after theme change, but expect heights have not changed
-      gridOptions = agGridService.getGridOptions(overrideOptions);
-      expect(gridOptions.rowHeight).toBe(60);
+      expect(agGridService.getHeaderHeight()).toBe(60);
     });
 
     it('should not overwrite default component definitions', () => {
@@ -976,69 +965,5 @@ describe('SkyAgGridService', () => {
     it('should not generate a class without an id', () => {
       expect(defaultGridOptions.getRowClass?.(params)).toBeFalsy();
     });
-  });
-});
-
-describe('SkyAgGridService via fixture', () => {
-  let gridWrapperFixture: ComponentFixture<SkyAgGridFixtureComponent>;
-  let mockThemeSvc: {
-    settingsChange: BehaviorSubject<SkyThemeSettingsChange>;
-  };
-
-  beforeEach(() => {
-    mockThemeSvc = {
-      settingsChange: new BehaviorSubject<SkyThemeSettingsChange>({
-        currentSettings: new SkyThemeSettings(
-          SkyTheme.presets.default,
-          SkyThemeMode.presets.light,
-        ),
-        previousSettings: undefined,
-      }),
-    };
-  });
-
-  it('should update header and row height via api', async () => {
-    TestBed.configureTestingModule({
-      imports: [SkyAgGridModule],
-      providers: [
-        {
-          provide: SkyThemeService,
-          useValue: mockThemeSvc,
-        },
-      ],
-    });
-    gridWrapperFixture = TestBed.createComponent(SkyAgGridFixtureComponent);
-    gridWrapperFixture.detectChanges();
-    await gridWrapperFixture.whenStable();
-
-    const api = gridWrapperFixture?.componentInstance?.agGrid?.api as GridApi;
-    const headerHeightSpy = spyOn(api, 'setHeaderHeight');
-    const rowHeightSpy = spyOn(api, 'resetRowHeights');
-    expect(api).toBeDefined();
-    expect(api.getSizesForCurrentTheme().headerHeight).toEqual(37);
-    expect(api.getSizesForCurrentTheme().rowHeight).toEqual(38);
-
-    mockThemeSvc.settingsChange.next({
-      currentSettings: new SkyThemeSettings(
-        SkyTheme.presets.modern,
-        SkyThemeMode.presets.light,
-      ),
-      previousSettings: undefined,
-    });
-    gridWrapperFixture.detectChanges();
-    await gridWrapperFixture.whenStable();
-    expect(headerHeightSpy).toHaveBeenCalledWith(60);
-    expect(rowHeightSpy).toHaveBeenCalled();
-
-    mockThemeSvc.settingsChange.next({
-      currentSettings: new SkyThemeSettings(
-        SkyTheme.presets.default,
-        SkyThemeMode.presets.light,
-      ),
-      previousSettings: undefined,
-    });
-    gridWrapperFixture.detectChanges();
-    await gridWrapperFixture.whenStable();
-    expect(headerHeightSpy).toHaveBeenCalledWith(37);
   });
 });