+++ libs/components/forms/src/lib/modules/file-attachment/file-attachment.component.ts
@@ -108,6 +108,14 @@ export class SkyFileAttachmentComponent
   public helpPopoverTitle: string | undefined;
 
   /**
+   * A help key that identifies the global help content to display. When specified, a [help inline](https://developer.blackbaud.com/skyux/components/help-inline) button is
+   * placed beside the single file attachment label. Clicking the button invokes global help as configured by the application.
+   * @preview
+   */
+  @Input()
+  public helpKey: string | undefined;
+
+  /**
    * The text to display as the file attachment's label.
    * @preview
    */